# demo/openapi.json
{"openapi":"3.0.1","info":{"title":"여행ON나 API","description":"여행ON나 애플리케이션의 API 문서입니다.","contact":{"name":"TravelONna Team","url":"https://travelonna.com","email":"contact@travelonna.com"},"license":{"name":"Apache License Version 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"},"version":"v1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"인증","description":"인증 관련 API"}],"paths":{"/api/auth/refresh":{"post":{"tags":["인증"],"summary":"토큰 갱신","description":"리프레시 토큰을 사용하여 액세스 토큰을 갱신합니다.","operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"토큰 갱신 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"400":{"description":"잘못된 요청","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"500":{"description":"서버 오류","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"유효하지 않은 리프레시 토큰","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/google":{"post":{"tags":["인증"],"summary":"Google 로그인","description":"Google OAuth2.0 인증 코드를 사용하여 로그인합니다.","operationId":"googleLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleTokenRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"500":{"description":"서버 오류","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"200":{"description":"로그인 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}}},"components":{"schemas":{"RefreshTokenRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}},"description":"리프레시 토큰"},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int32"}}},"GoogleTokenRequest":{"required":["code"],"type":"object","properties":{"code":{"type":"string"}},"description":"Google 인증 코드"}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}