# demo/src/main/java/com/travelonna/demo/domain/auth/dto/GoogleTokenRequest.java
package com.travelonna.demo.domain.auth.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@NoArgsConstructor
@AllArgsConstructor
public class GoogleTokenRequest {
    
    @NotBlank(message = "Authorization code is required")
    private String code;
} 